requirements.txt
django
djangorestframework
django_extensions
python-decouple
Pillow
python-magic-bin
djangorestframework-simplejwt

